Binomial implemetation

+++ Binomial.java
@@ -12,7 +12,7 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, basic version.
-	public static long binomial1(int n, int k) { 
+	public static int binomial1(int n, int k) { 
 		if (k > n) return 0;
 		if (k == 0 || n == 0) return 1;
 		return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
@@ -20,7 +20,7 @@ public class Binomial {
 	 }
 	
 	// Computes the Binomial function, efficiently
-	public static long binomial(int n, int k) {
+	public static int binomial(int n, int k) {
 		//// This function creates a 2D array, say memo, 
 		// and then initializes all its elements to -1.
 		//// It then calls binomial(n, k, memo), which does all the heavy lifiting.
@@ -38,7 +38,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static long binomial(int n, int k, long[][] memo) {
+	private static int binomial(int n, int k, int[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
